+++ src/main/java/webserver/RequestHandler.java
@@ -7,18 +7,16 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.net.Socket;
-import java.net.URISyntaxException;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
 import controller.home.HomeController;
-import controller.auth.LogInController;
-import controller.auth.SignUpController;
+import controller.user.LogInController;
+import controller.user.SignUpController;
 import controller.user.UserListController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import utils.FileIoUtils;
 import utils.IOUtils;
 import webserver.http.request.HttpRequest;
 import webserver.http.request.RequestBody;
